- move apiIntegrationTestUtil to __integration__ - ignore __integration__ in .npmignore and .mailing build step
